requirements.txt
streamlit==1.44.1
pandas==2.2.3
python-dotenv==1.1.0
langchain==0.3.23
langchain-core==0.3.52
langchain-openai==0.3.13
langchain-ollama==0.3.1
langchain-groq==0.3.2
langchain-google-genai==2.1.2
langchain-deepseek==0.1.3
tqdm==4.67.1
beautifulsoup4==4.13.4
requests==2.32.3
pypdf==5.4.0
html2text==2025.4.15